requirements.txt
alabaster==0.7.12
asgiref==3.2.3
astroid==2.3.1
autopep8==1.5
Babel==2.8.0
boto3==1.12.25
botocore==1.15.25
certifi==2020.4.5.1
chardet==3.0.4
dateparser==0.7.4 
dj-database-url==0.5.0
Django==3.0.3
django-appconf==1.0.4
django-cors-headers==3.2.1
django-filter==2.2.0
django-imagekit==4.0.2
django-select2==7.2.3
django-storages==1.9.1
django-suit2==2.0a1
djangorestframework==3.11.0
docutils==0.15.2
google-api-python-client==1.4.1
gunicorn==20.0.4
httplib2==0.17.3
idna==2.9
isort==4.3.21
jet-bridge-base==0.7.6
jet-django==1.0.1
Jinja2==2.11.2
jmespath==0.9.5
lazy-object-proxy==1.4.2
Markdown==3.2.1
MarkupSafe==1.1.1
mccabe==0.6.1
oauth2client==4.1.3
pilkit==2.0
Pillow==7.0.0
prompt-toolkit==2.0.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
Pygments==2.6.1
pylint==2.4.2
python-dateutil==2.8.1
pytz==2019.3
r-django-jet==2.0.8
regex==2020.4.4
requests==2.23.0
rsa==4.0
s3transfer==0.3.3
six==1.12.0
snowballstemmer==2.0.0
Sphinx==1.3.5
sphinx-rtd-theme==0.1.9
SQLAlchemy==1.3.16
sqlparse==0.3.1
typed-ast==1.4.0
tzlocal==2.0.0
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.9
whitenoise==5.0.1
wrapt==1.11.2
